+++ native~/Runtime/src/CesiumMetadataImpl.cpp
@@ -52,6 +52,11 @@ int64_t getFeatureIdFromVertex(
 
 } // namespace
 
+CesiumMetadataImpl::CesiumMetadataImpl(
+    const DotNet::CesiumForUnity::CesiumMetadata& metadata) {}
+
+CesiumMetadataImpl::~CesiumMetadataImpl() {}
+
 void CesiumMetadataImpl::addMetadata(
     int32_t instanceID,
     const CesiumGltf::Model* pModel,